% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIGHR.R
\name{SIGHR}
\alias{SIGHR}
\title{A SIGHR Function}
\usage{
SIGHR(X, Y, D, Dtil, q, coef, prob, iter)
}
\arguments{
\item{X:}{the design matrix with confounding variables, SNPs.}

\item{Y:}{the response variable.}

\item{D:}{the auxiliary matrix.}

\item{Dtil:}{list of CMP infor. 1. loc: each row is a location selected from the auxiliary information space. 2. hyper: each row is a set of hyperparameteres for the corresponding location.}

\item{q:}{the number of confounding (sex, age, BMI... including intercept)}

\item{coef:}{the intial value of (alpha, beta)}

\item{prob:}{the intial value inclusion values of beta, of length (p-q)}
}
\value{
an MCMC sample of all regression coefficients ("coef") including the ones for the intercept and confounders in (1), indicators for all beta's ("indicator.z"), all regression coefficients in the hierarchical logistic regression ("gamma"), and the variance ("sig2") in (1).
}
\description{
This function allows to leverage side information from other studies to inform about the sparsity structure of the regression coefficients.
}
